Cleaned up all class, enum, and viewmodel files
