+++ BackChannel/BackChannel/ViewModels/ChatViewModel.cs
@@ -1,13 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
-using System.Linq;
 using System.Runtime.CompilerServices;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
-using System.Windows.Controls;
 using BackChannel.Classes;
 
 namespace BackChannel.ViewModels
@@ -17,7 +12,14 @@ namespace BackChannel.ViewModels
     /// </summary>
     public class ChatViewModel : INotifyPropertyChanged
     {
+        // UI Notification functions
         public event PropertyChangedEventHandler PropertyChanged;
+        protected void OnPropertyChanged([CallerMemberName] string name = null)
+        {
+            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(name));
+        }
+
+        // Message collection
         private ObservableCollection<Message> messages = new ObservableCollection<Message>();
         public ObservableCollection<Message> Messages
         {
@@ -28,11 +30,8 @@ namespace BackChannel.ViewModels
                 this.OnPropertyChanged("Messages");
             }
         }
-        protected void OnPropertyChanged([CallerMemberName] string name = null)
-        {
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(name));
-        }
 
+        // Helper functions
         public void AddText(Message t)
         {
             Application.Current.Dispatcher.Invoke(new Action(() =>
@@ -40,7 +39,6 @@ namespace BackChannel.ViewModels
                 Messages.Add(t);
             }));
         }
-
         public void RemoveText(Message t)
         {
             Application.Current.Dispatcher.Invoke(new Action(() =>
